Avoid potential overflow in int multiplication before it is converted to long.
(cherry picked from commit b43f5b8cf28a8fb8e3418536f0a45af5c5d526e4)

+++ activemq-client/src/main/java/org/apache/activemq/util/BitArrayBin.java
@@ -175,7 +175,7 @@ public class BitArrayBin implements Serializable {
                 last = list.get(lastBitArrayIndex);
                 if (last != null) {
                     result += last.length() -1;
-                    result += lastBitArrayIndex * BitArray.LONG_SIZE;
+                    result += lastBitArrayIndex * (long) BitArray.LONG_SIZE;
                     break;
                 }
             }
